edit, delete post only owner

+++ backend/routes/posts.js
@@ -37,6 +37,7 @@ router.post(
     title: req.body.title,
     content: req.body.content,
     imagePath: url + '/images/' + req.file.filename,
+    creator: req.userData.userId,
   });
   post.save().then(createdPost => {
     res.status(201).json({
@@ -64,9 +65,14 @@ router.put(
     title: req.body.title,
     content: req.body.connect,
     imagePath: imagePath,
+    creator: req.userData.userId,
   });
-  Post.updateOne({_id: req.params.id}, post).then(result => {
-    res.status(200).json({ message: 'Update succesful!' });
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result => {
+    if (result.nModified > 0) {
+      res.status(200).json({ message: 'Update succesful!' });
+    } else {
+      res.status(401).json({ message: 'Not authorized' });
+    }
   });
 });
 
@@ -108,11 +114,16 @@ router.delete(
   '/:id',
   checkAuth,
   (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id })
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId })
   .then((result) => {
-    res.status(200).json({
-      message: 'Post deleted'
-    });
+    if (result.n > 0) {
+      res.status(200).json({ message: 'Delete succesful!' });
+    } else {
+      res.status(401).json({ message: 'Not authorized' });
+    }
+  })
+  .catch(err => {
+    console.log(err);
   });
 });
